WL#4389 Transform EXISTS subqueries to semi-join
Extend semi-join check to accept EXISTS subqueries in addition to IN.
Filter out non-deterministic subqueries.

Decorrelate subquery WHERE condition:

Decorrelation in this case means decomposing a WHERE condition consisting
of equality predicates between an expression with a local table reference
and an expression with a table reference from the immediately outer
query block. Other predicates cannot be decorrelated and are kept in the
subquery WHERE condition.

The decorrelated expressions are added to sj_outer_exprs and sj_inner_exprs,
alongside the expressions from the IN predicate (not applicable for EXISTS).

After decorrelation, a full semi-join condition is synthesized from
sj_inner_exprs and sj_outer_exprs, in addition to the remainder of the
subquery's WHERE condition.

In addition, sj_corr_tables (the set of outer non-trivially correlated tables)
is calculated a bit differently: after decorrelation, it is the set of
tables from the WHERE condition remainder originating from the outer query
block (notice it is calculated after fix_after_pullout() is performed).
Likewise, sj_depends_on is calculated after predicate synthesizing as the
set of tables from the outer query block (before semi-join transformation).

Test changes:
- Several EXISTS subqueries are transformed to semi-join.
- subquery_sj.inc is extended with a similar EXISTS subquery.
- subquery_exists.test is a new test file for testing EXISTS queries.

+++ sql/sql_planner.cc
@@ -883,7 +883,7 @@ double Optimize_table_order::lateral_derived_cost(
   table_map deps = tab->table_ref->derived_unit()->m_lateral_deps;
   POSITION *positions = got_final_plan ? join->best_positions : join->positions;
   double derived_mat_cost = 0;
-  for (uint j = idx; j >= join->const_tables; j--) {
+  for (int j = idx; j >= (int)join->const_tables; j--) {
     if (deps & join->best_ref[j]->table_ref->map()) {
       // We found the last table in plan, on which 'tab' depends.
       auto res = tab->table_ref->derived_unit()->query_result();
